requirements.txt
fastapi
uvicorn
pydantic

langchain
langchain-community
langchain-huggingface

faiss-cpu

sentence-transformers
transformers
torch

gitpython

python-dotenv
tqdm

python-multipart
chardet
pydantic